[ActionRecognition,SpecializedAction,Action] set the ttl of the buffer at the max of ttl of all actions

+++ src/core/Types/SpecializedAction.cpp
@@ -12,7 +12,7 @@ SpecializedAction::SpecializedAction(const std::string& name,
                                      const std::vector<procedural::PatternTransitionFact_t>& patterns,
                                      const std::vector<PatternTransitionNetwork_t>& patterns_network,
                                      const std::vector<ActionDescription_t>& descriptions,
-                                     uint32_t ttl) : name_(name),
+                                     double ttl) : name_(name),
                                                      is_valid_(false),
                                                      time_to_live_(ttl),
                                                      id_(0)
